+++ src/gen_michelson.ml
@@ -869,6 +869,7 @@ let to_ir (model : M.model) : T.ir =
     | Mshiftleft (l, r)  -> T.Ibinop (Blsl, f l, f r)
     | Mshiftright (l, r) -> T.Ibinop (Blsr, f l, f r)
     | Msubnat (l, r)     -> let b = T.Bsubnat in add_builtin b; T.Icall (get_fun_name b, [f l; f r], is_inline b)
+    | Msubmutez (l, r)   -> T.isub_mutez (f l) (f r)
 
 
     (* asset api effect *)