+++ src/gen_model.ml
@@ -757,6 +757,11 @@ let to_model (ast : A.ast) : M.model =
         let fy = f y in
         M.Msubnat (fx, fy)
 
+      | A.Pcall (None, A.Cconst A.Csubmutez, [AExpr x; AExpr y]) ->
+        let fx = f x in
+        let fy = f y in
+        M.Msubmutez (fx, fy)
+
       | A.Pcall (None, A.Cconst A.Cblake2b, [AExpr x]) ->
         let fx = f x in
         M.Mblake2b (fx)