+++ src/model.ml
@@ -308,6 +308,7 @@ type ('id, 'term) mterm_node  =
   | Mshiftleft        of 'term * 'term
   | Mshiftright       of 'term * 'term
   | Msubnat           of 'term * 'term
+  | Msubmutez         of 'term * 'term
   (* asset api effect *)
   | Maddasset         of ident * 'term
   | Maddfield         of ident * ident * 'term * 'term (* asset_name * field_name * asset instance * item *)
@@ -1521,6 +1522,7 @@ let cmp_mterm_node
     | Mshiftleft (l1, r1), Mshiftleft (l2, r2)                                         -> cmp l1 l2 && cmp r1 r2
     | Mshiftright (l1, r1), Mshiftright (l2, r2)                                       -> cmp l1 l2 && cmp r1 r2
     | Msubnat (l1, r1), Msubnat (l2, r2)                                               -> cmp l1 l2 && cmp r1 r2
+    | Msubmutez (l1, r1), Msubmutez (l2, r2)                                           -> cmp l1 l2 && cmp r1 r2
     (* asset api effect *)
     | Maddasset (an1, i1), Maddasset (an2, i2)                                         -> cmp_ident an1 an2 && cmp i1 i2
     | Maddfield (an1, fn1, c1, i1), Maddfield (an2, fn2, c2, i2)                       -> cmp_ident an1 an2 && cmp_ident fn1 fn2 && cmp c1 c2 && cmp i1 i2
@@ -1966,6 +1968,7 @@ let map_term_node_internal (fi : ident -> ident) (g : 'id -> 'id) (ft : type_ ->
   | Mshiftleft (l, r)              -> Mshiftleft (f l, f r)
   | Mshiftright (l, r)             -> Mshiftright (f l, f r)
   | Msubnat (l, r)                 -> Msubnat (f l, f r)
+  | Msubmutez (l, r)               -> Msubmutez (f l, f r)
   (* asset api effect *)
   | Maddasset (an, i)              -> Maddasset (fi an, f i)
   | Maddfield (an, fn, c, i)       -> Maddfield (fi an, fi fn, f c, f i)
@@ -2408,6 +2411,7 @@ let fold_term (f : 'a -> ('id mterm_gen) -> 'a) (accu : 'a) (term : 'id mterm_ge
   | Mshiftleft (l, r)                     -> f (f accu l) r
   | Mshiftright (l, r)                    -> f (f accu l) r
   | Msubnat (l, r)                        -> f (f accu l) r
+  | Msubmutez (l, r)                      -> f (f accu l) r
   (* asset api effect *)
   | Maddasset (_, i)                      -> f accu i
   | Maddfield (_, _, c, i)                -> f (f accu c) i
@@ -3149,6 +3153,11 @@ let fold_map_term
     let le, la = f accu l in
     let re, ra = f la r in
     g (Msubnat (le, re)), ra
+
+  | Msubmutez (l, r) ->
+    let le, la = f accu l in
+    let re, ra = f la r in
+    g (Msubmutez (le, re)), ra
 
 
   (* asset api effect *)