+++ src/printer_model.ml
@@ -783,6 +783,14 @@ let pp_mterm fmt (mt : mterm) =
       in
       pp fmt (l, r)
 
+    | Msubmutez (l, r) ->
+      let pp fmt (l, r) =
+        Format.fprintf fmt "sub_mutez(%a, %a)"
+          f l
+          f r
+      in
+      pp fmt (l, r)
+
 
     (* asset api effect *)