+++ src/typing.ml
@@ -1427,7 +1427,8 @@ let cryptoops : opinfo list =
 (* -------------------------------------------------------------------- *)
 let mathops : opinfo list =
   [
-    op "sub_nat" A.Csubnat `Partial (Some A.vtnat) [ A.vtnat ] (`Ty A.vtnat) Mint.empty
+    op "sub_nat" A.Csubnat `Partial (Some A.vtnat) [ A.vtnat ] (`Ty A.vtnat) Mint.empty;
+    op "sub_mutez" A.Csubmutez `Total (Some A.vtcurrency) [ A.vtcurrency ] (`Ty (A.Toption A.vtcurrency)) Mint.empty
   ]
 
 (* -------------------------------------------------------------------- *)